refactor create request to be in requestActions module instead of request form component
